UTU/requirements.txt
cachelib==0.1.1
click==7.1.2
dnspython==2.1.0
Flask==1.1.2
Flask-Session==0.3.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongo==0.2.0
numpy==1.19.5
pandas==1.2.0
pymongo==3.11.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
six==1.15.0
Werkzeug==1.0.1
